+++ src/Ext/WeaponType/Body.h
@@ -30,6 +30,8 @@ public:
 		Valueable<AffectedHouse> CanTargetHouses;
 		ValueableVector<int> Burst_Delays;
 		Valueable<AreaFireTarget> AreaFire_Target;
+		Valueable<bool> DetachedFromOwner;
+		Nullable<WeaponTypeClass*> FeedbackWeapon;
 
 		ExtData(WeaponTypeClass* OwnerObject) : Extension<WeaponTypeClass>(OwnerObject)
 			, DiskLaser_Radius { 38.2 }
@@ -45,6 +47,8 @@ public:
 			, CanTargetHouses { AffectedHouse::All }
 			, Burst_Delays {}
 			, AreaFire_Target { AreaFireTarget::Base }
+			, DetachedFromOwner { false }
+			, FeedbackWeapon {}
 		{ }
 
 		virtual ~ExtData() = default;